Correct bug in the remove function
If the argument in the remove function has a match with head and if the list length is 1 then the return value of head of the new list should be "null" (same as in the start of the linkedList when it did not have any entry) but this was returning undefined. So I corrected that.